feat: timeout limit added to arguments, updated example
